+++ CHANGELOG.md
@@ -11,12 +11,14 @@ SPDX-License-Identifier: MIT
 05-08-2026
 
 - Added two new components `BreadcrumbNav` & `RemoteBreadcrumbNav`
+- Add advanced translations to Zones
+- Add filter exclude incoming to Zones and Components
 
 ## tol-ui `4.0.7`
 
 30-07-2026
 
-- Added footer content override for SmartApp
+- Added footer content override for `SmartApp`
 
 ## tol-ui `4.0.6`